+++ blueprints/routes.py
@@ -16,12 +16,10 @@ import re
 import nltk
 from nltk.tokenize import word_tokenize
 from nltk.corpus import stopwords
-from controllers.pdfReader import compare_resume_to_metrics, extract_keywords_from_metrics, extract_text_from_pdf, preprocess_text
+from controllers.resume import  compare_resume_to_metrics, extract_keywords_from_metrics, extract_text_from_pdf, preprocess_text
 from controllers.dumpText import insert_resume_text
 from pathlib import Path
 
-
-import time
 
 home_blueprint = Blueprint('home', __name__)
 login_blueprint = Blueprint('login', __name__)
