+++ blueprints/templates/profile.html
@@ -51,7 +51,6 @@
         {% endif %}
     </div>
 </div>
-
 <br><br><br><br><br><br><br>
 <div>
 <h1>Upload your CV for Feedback</h1>
@@ -69,4 +68,40 @@
         {% endif %}
     {% endwith %}
 </div>
+
+<!-- New form for additional request -->
+<div>
+<h1>Update User's Skills</h1>
+    <form method="POST" action="/update_user_skills" enctype="multipart/form-data">
+            <label for="Resume-File" >File Upload*</span></label>
+            <input type="file" class="form-control" name="Resume-File" access="false" multiple="false" id="Resume-File" required="required" aria-required="true">
+            
+            <label for="highest_education_obtained" class="formbuilder-select-label">Highest Education Obtained*<span class="tooltip-element" tooltip="Degree">?</span></label>
+            <select class="form-control" name="highest_education_obtained" id="highest_education_obtained" required="required" aria-required="true">
+            <option value="diploma" selected="true" id="highest_education_obtained-0">Diploma</option>
+            <option value="advance_diploma" id="highest_education_obtained-1">Advance Diploma</option>
+            <option value="bachelors_degree" id="highest_education_obtained-2">Bachelor's Degree</option>
+            <option value="master_degree" id="highest_education_obtained-3">Master's Degree</option>
+            <option value="phd" id="highest_education_obtained-4">Phd</option>
+            </select>
+
+            <label for="number_of_hours" >Number of Hours in the Field of Interest *<span class="tooltip-element" tooltip="0 - 10000 HRS">?</span></label>
+            <input type="text" placeholder="1234" class="form-control" name="number_of_hours" access="false" id="number_of_hours" title="0 - 10000 HRS" required="required" aria-required="true">
+
+            <label for="certification_number" >Number of Certification in the Field of Interest.<span class="tooltip-element" tooltip="Example: User ABC has AWS Cloud Practitioner and ECC CEH.  User has 2 certifications, hence the input will be 2">?</span></label>
+            <input type="text" placeholder="1" class="form-control" name="certification_number" access="false" id="certification_number" title="Example: User ABC has AWS Cloud Practitioner and ECC CEH.  User has 2 certifications, hence the input will be 2" required="required" aria-required="true">
+ 
+            <button type="submit">Upload</button>
+    </form>
+        {% with messages = get_flashed_messages() %}
+            {% if messages %}
+                <ul>
+                {% for message in messages %}
+                    <li>{{ message }}</li>
+                {% endfor %}
+                </ul>
+            {% endif %}
+        {% endwith %}
+
+
 {% endblock %}
