--- deleteUser.sql
@@ -1,2 +1,0 @@
--- SQLite
-DELETE FROM userdata WHERE id = 1
